requirements.txt
anyio==3.7.0
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastapi==0.97.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.3
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.9
Pygments==2.15.1
pytest==7.3.2
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
wcwidth==0.2.6
